requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
google-generativeai==0.7.2
# faster-whisper==1.0.3
gtts==2.5.1
groq==0.11.0
pymupdf==1.24.5
python-dotenv==1.0.1
httpx==0.27.0
motor==3.3.2
pymongo==4.6.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
email-validator==2.1.1

# fastapi==0.111.0
# uvicorn[standard]==0.30.1
# python-multipart==0.0.9
# google-generativeai==0.7.2
# # faster-whisper==1.0.3
# gtts==2.5.1
# groq==0.11.0
# pymupdf==1.24.5
# python-dotenv==1.0.1
# httpx==0.27.0